+++ package.json
@@ -30,10 +30,6 @@
     "test_eps": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_eps.py",
     "test_eps_hk_only": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_eps_hk_only.py",
     "test_dfgm": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_dfgm.py",
-    "test_iris": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_iris.py",
-    "refactor:cli": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 src/refactor/cli.py",
-    "refactor:sat_cli": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 src/refactor/sat_cli.py",
-    "refactor:sat_update": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 src/refactor/updater.py",
-    "refactor:ftp": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 src/refactor/ftp.py"
+    "test_iris": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_iris.py"
   }
 }
