--- src/groundStation/__init__.py
@@ -1,6 +1,0 @@
-try:
-    from ex2_ground_station_software.src.groundStation import groundStation
-    from ex2_ground_station_software.src.groundStation import cliGroundStation
-    from ex2_ground_station_software.src.groundStation import updater
-except ModuleNotFoundError:
-    pass
